config: Consolidate parser and writer in same file
